third_party/WebKit/LayoutTests/fast/events/drag-and-drop-set-drag-data-arguments.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<div draggable="true" id="dragger">Drag me</div>
<script>

description('Tests correct behavior of event.dataTransfer.setData with wrong number of arguments');

var event;

var dragger = document.getElementById('dragger');
dragger.addEventListener('dragstart', function(e) {
    event = e;
    shouldThrow('event.dataTransfer.setData()', '"TypeError: Failed to execute \'setData\' on \'DataTransfer\': 2 arguments required, but only 0 present."');
    shouldThrow('event.dataTransfer.setData("Text")', '"TypeError: Failed to execute \'setData\' on \'DataTransfer\': 2 arguments required, but only 1 present."');
}, false);

if (window.eventSender) {
    var rect = dragger.getBoundingClientRect();
    var x = rect.left + rect.width / 2;
    var y = rect.top + rect.height / 2;
    eventSender.mouseMoveTo(x, y);
    eventSender.mouseDown();
    eventSender.leapForward(400);
    eventSender.mouseMoveTo(x + 100, y);
    eventSender.mouseUp();
}

</script>
